project/pyproject.toml
[project]
name = "project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "minsearch>=0.0.7",
    "openai>=2.14.0",
    "python-frontmatter>=1.1.0",
    "requests>=2.32.5",
]
